+++ src/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.6 (2023-10-13)
+
+### Fixed
+ - Fix issue with sending commands
+
+ ### Updated
+ - Add state tracking
+
 ## 1.0.5 (2023-10-11)
 
 ### Updated